refactor(review-changes): reorganize component structure and remove unused files
- Updated import path for ReviewChangesPage component.
- Deleted unused ReviewChangesDetailsScreen, ReviewChangesPage, ReviewChangesScreen, and ReviewCollectionGrid components along with their associated stories.
- Streamlined the codebase by removing obsolete files to enhance maintainability.
